add falsely released migration
this migration was released on aldryn in version 0.2.4, even though
it should not have (it was actually in another branch with yet
unreleased changes). Since we already have installations of this
Addon in the wild, that have already migrated this migration, I’m
adding it. This avoids “ghost migration” errors.
Since it’s a harmless migration this should not cause any problems.