namespace: do not delete jobcategory_new, remove language_code changes
We'll leave columns/tables for migrating to aldryn-categories, so that next if it was used somewhere we can grab data from them.

Also changes for language_code that was picked up by django 1.7 migrations are delted here. They should be either handled by parler, or would be added later with sepparate migration. Most likely adding not related to PR chagnes is not the best idea.