Fix django 1.7 cms 3.1 migration 0004 to use appropriate model.
There was an issue because of migrations method "apps.get_model" picked up legacy db schema (cms 0003 which is pointed as a dependency) which caused a migration error if cms migrations were performed already. Though on a fresh database if all dependencies/packages are installed and configured it would work since migrations are applied in the right order.